# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: CoffeeChain Trace API
  description: |
    REST API for the CoffeeChain coffee supply-chain traceability backend.
    All authenticated endpoints require a Bearer JWT obtained from POST /api/auth/login.
  version: 1.0.0
  contact:
    name: CoffeeChain Backend Unit-2
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/roast:
    post:
      tags: [Roaster]
      summary: Create roast batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoastBatchRequest"
      responses:
        "200":
          description: Created batch from ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/roast/{id}/evidence:
    post:
      tags: [Roaster]
      summary: Add IPFS evidence to a roast batch
      parameters:
        - $ref: "#/components/parameters/batchId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddEvidenceRequest"
      responses:
        "200":
          description: Updated batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchView"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/roast/{id}/status:
    patch:
      tags: [Roaster]
      summary: Update roast batch status
      parameters:
        - $ref: "#/components/parameters/batchId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Updated batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/transfer/request:
    post:
      tags: [Roaster]
      summary: "Request transfer of roast batch to Org2 (body: { batchId, toMSP })"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
      responses:
        "200":
          description: Batch with TRANSFER_PENDING status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchView"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─────────────────────────────────────────────
  # Packager
  # ─────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    batchId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Internal batch UUID on the ledger
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but insufficient role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
...
    BatchView:
      type: object
      additionalProperties: true
      description: Raw JSON returned from the Fabric chaincode (schema mirrors Batch.java on ledger)
...
    CreateRoastBatchRequest:
      type: object
      required:
        [parentBatchId, roastProfile, roastDate, roastDurationMinutes, weightKg]
      properties:
        parentBatchId:
          type: string
        roastProfile:
          type: string
        roastDate:
          type: string
          format: date
        roastDurationMinutes:
          type: string
        weightKg:
          type: string
...
    AddEvidenceRequest:
      type: object
      required: [evidenceHash, evidenceUri]
      properties:
        evidenceHash:
          type: string
        evidenceUri:
          type: string
    TransferRequest:
      type: object
      required: [batchId, toMSP]
      properties:
        batchId:
          type: string
          description: ID of the batch to transfer
        toMSP:
          type: string
          enum: [Org2MSP]
    UpdateStatusRequest:
      type: object
      required: [newStatus]
      properties:
        newStatus:
          type: string
          enum:
            [
              CREATED,
              IN_PROCESS,
              COMPLETED,
              TRANSFER_PENDING,
              TRANSFERRED,
              IN_STOCK,
              SOLD,
            ]
        finalWeightKg:
          type: string
          description: Final weight (required when newStatus=COMPLETED for harvest/processed/roast)
        roastDurationMinutes:
          type: string
          description: Roast duration in minutes (required when newStatus=COMPLETED for roast)